Migrate upsert to create record design
