+++ packages/twenty-front/src/modules/workflow/workflow-steps/workflow-actions/components/WorkflowEditActionUpsertRecord.tsx
@@ -1,11 +1,13 @@
 import { SidePanelHeader } from '@/command-menu/components/SidePanelHeader';
 import { type WorkflowUpsertRecordAction } from '@/workflow/types/Workflow';
 import { WorkflowStepFooter } from '@/workflow/workflow-steps/components/WorkflowStepFooter';
-import { WorkflowUpdateRecordBody } from '@/workflow/workflow-steps/workflow-actions/components/WorkflowUpdateRecordBody';
+import {
+  WorkflowCreateRecordBody,
+  type CreateRecordFormData,
+} from '@/workflow/workflow-steps/workflow-actions/components/WorkflowCreateRecordBody';
 
 import { UPSERT_RECORD_ACTION } from '@/workflow/workflow-steps/workflow-actions/constants/actions/UpsertRecordAction';
 import { useWorkflowActionHeader } from '@/workflow/workflow-steps/workflow-actions/hooks/useWorkflowActionHeader';
-import { type UpdateRecordFormData } from '@/workflow/workflow-steps/workflow-actions/types/update-record-form-data.type';
 import { useIcons } from 'twenty-ui/display';
 
 type WorkflowEditActionUpsertRecordProps = {
@@ -34,19 +36,20 @@ export const WorkflowEditActionUpsertRecord = ({
 
   const isFormDisabled = actionOptions.readonly === true;
 
-  const handleUpdate = (formData: UpdateRecordFormData) => {
+  const handleUpdate = (formData: CreateRecordFormData) => {
     if (actionOptions.readonly === true) {
       return;
     }
+
+    const { objectName: updatedObjectName, ...updatedOtherFields } = formData;
 
     actionOptions.onActionUpdate({
       ...action,
       settings: {
         ...action.settings,
         input: {
-          objectName: formData.objectNameSingular,
-          objectRecord: formData,
-          fieldsToUpdate: formData.fieldsToUpdate,
+          objectName: updatedObjectName,
+          objectRecord: updatedOtherFields,
         },
       },
     });
@@ -72,14 +75,12 @@ export const WorkflowEditActionUpsertRecord = ({
         disabled={isFormDisabled}
         iconTooltip={UPSERT_RECORD_ACTION.defaultLabel}
       />
-      <WorkflowUpdateRecordBody
-        defaultObjectNameSingular={action.settings.input.objectName}
-        defaultFieldsToUpdate={action.settings.input.fieldsToUpdate}
+      <WorkflowCreateRecordBody
+        defaultObjectName={action.settings.input.objectName}
         defaultObjectRecord={action.settings.input.objectRecord}
         readonly={isFormDisabled}
         actionType="UPSERT_RECORD"
         onUpdate={handleUpdate}
-        shouldPickRecord={false}
       />
       {!actionOptions.readonly && <WorkflowStepFooter stepId={action.id} />}
     </>
