+++ packages/twenty-front/src/modules/workflow/workflow-steps/workflow-actions/utils/shouldDisplayFormField.ts
@@ -3,7 +3,7 @@ import { type WorkflowActionType } from '@/workflow/types/Workflow';
 import { CustomError } from 'twenty-shared/utils';
 import { FieldMetadataType } from '~/generated-metadata/graphql';
 
-const COMMON_DISPLAYABLE_FIELD_TYPES = [
+const SUPPORTED_FORM_FIELD_TYPES = [
   FieldMetadataType.TEXT,
   FieldMetadataType.NUMBER,
   FieldMetadataType.DATE,
@@ -19,10 +19,6 @@ const COMMON_DISPLAYABLE_FIELD_TYPES = [
   FieldMetadataType.DATE_TIME,
   FieldMetadataType.RAW_JSON,
   FieldMetadataType.UUID,
-];
-
-const FIND_RECORDS_DISPLAYABLE_FIELD_TYPES = [
-  ...COMMON_DISPLAYABLE_FIELD_TYPES,
   FieldMetadataType.ARRAY,
   FieldMetadataType.RELATION,
 ];
@@ -34,35 +30,28 @@ export const shouldDisplayFormField = ({
   fieldMetadataItem: FieldMetadataItem;
   actionType: WorkflowActionType;
 }) => {
-  let isTypeAllowedForAction = false;
+  if (!SUPPORTED_FORM_FIELD_TYPES.includes(fieldMetadataItem.type)) {
+    return false;
+  }
+
   const isIdField = fieldMetadataItem.name === 'id';
+  const isNotSupportedRelation =
+    fieldMetadataItem.type === FieldMetadataType.RELATION &&
+    fieldMetadataItem.settings?.['relationType'] !== 'MANY_TO_ONE';
 
   switch (actionType) {
     case 'CREATE_RECORD':
-      isTypeAllowedForAction =
-        fieldMetadataItem.type !== FieldMetadataType.RELATION ||
-        fieldMetadataItem.settings?.['relationType'] === 'MANY_TO_ONE';
-      return (
-        isTypeAllowedForAction &&
-        !fieldMetadataItem.isSystem &&
-        fieldMetadataItem.isActive
-      );
     case 'UPDATE_RECORD':
     case 'UPSERT_RECORD':
-      isTypeAllowedForAction =
-        COMMON_DISPLAYABLE_FIELD_TYPES.includes(fieldMetadataItem.type) ||
-        fieldMetadataItem.settings?.['relationType'] === 'MANY_TO_ONE';
       return (
-        isTypeAllowedForAction &&
+        !isNotSupportedRelation &&
+        !fieldMetadataItem.isUIReadOnly &&
         !fieldMetadataItem.isSystem &&
         fieldMetadataItem.isActive
       );
     case 'FIND_RECORDS':
-      isTypeAllowedForAction = FIND_RECORDS_DISPLAYABLE_FIELD_TYPES.includes(
-        fieldMetadataItem.type,
-      );
       return (
-        isTypeAllowedForAction &&
+        !isNotSupportedRelation &&
         (!fieldMetadataItem.isSystem || isIdField) &&
         fieldMetadataItem.isActive
       );
