+++ packages/twenty-server/src/engine/core-modules/record-crud/types/record-crud-input.type.ts
@@ -43,5 +43,4 @@ export type FindRecordsInput = {
 export type UpsertRecordInput = {
   objectName: string;
   objectRecord: ObjectRecordProperties;
-  fieldsToUpdate?: string[];
 };
