+++ packages/twenty-server/src/modules/workflow/workflow-executor/workflow-actions/record-crud/upsert-record.workflow-action.ts
@@ -74,7 +74,6 @@ export class UpsertRecordWorkflowAction implements WorkflowAction {
     const toolOutput = await this.upsertRecordService.execute({
       objectName: workflowActionInput.objectName,
       objectRecord: workflowActionInput.objectRecord,
-      fieldsToUpdate: workflowActionInput.fieldsToUpdate,
       workspaceId,
       rolePermissionConfig: executionContext.rolePermissionConfig,
     });
